Added non-65C02 variant of test program.
The C-only code to access the W5100 used byte pointers to access the W5100 I/O ports. This has the benefit of allowing to set the I/O port address at run time - and thus allowing to set the Apple II slot the Uthernet II is in.

However the cc65 optmizer isn't able to generate code to properly access the W5100 I/O ports (which have side effects for each read !) without using 65C02 opcodes. So in order to come up with a test program variant that runs on non-65C02 machines the byte pointers had to be replaced with numeric constants casted to byte pointers. cc65 has a special optimization that transforms access to those casted constants in straight "LDA/STA <numeric constant>" opcodes. The obvious downside is that the Apple II slot the Uthernet II is in has to be set at compile time. Here it is slot 3, hence the name of the new files.